add font components
